requirements.txt
APScheduler==3.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
cachetools==4.2.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
et-xmlfile==1.1.0
exceptiongroup==1.1.0
frozenlist==1.3.3
google-ai-generativelanguage==0.3.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-translate==3.12.0
google-generativeai==0.2.0
googleapis-common-protos==1.59.0
h11==0.14.0
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
numpy==1.23.5
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.2
pandas-stubs==1.5.2.221124
prompt-toolkit==3.0.38
pycparser==2.21
Pygments==2.15.1
PySocks==1.7.1
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
rich==13.4.2
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
style==1.1.0
telegram==0.0.1
tornado==6.3.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
types-pytz==2022.6.0.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
update==0.0.1
urllib3==1.26.13
wcwidth==0.2.6
wsproto==1.2.0
yarl==1.9.2
aiogram==2.25.1
